requirements.txt
google-genai>=0.6.0
pydantic>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
jupyter>=1.0.0 